==> 8_Merging_DataFrames.ipynb <==
# Importing the desired packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)# To see all the columns of a dataframe
pd.set_option('display.max_rows', None)

# Function to reduce the memory usage of various Dataframes
def reduce_mem_usage(df):
    """ iterate through all the columns of a dataframe and modify the data type
        to reduce memory usage.        
       
        1. Iterate over every column
        2. Determine if the column is numeric
        3. Determine if the column can be represented by an integer
        4. Find the min and the max value
        5. Determine and apply the smallest datatype that can fit the range of values

    """
    start_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage of dataframe is {:.2f} MB'.format(start_mem))
    
    for col in df.columns:
        col_type = df[col].dtype
        
        if col_type != object:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
        else:
            df[col] = df[col].astype('category')

    end_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage after optimization is: {:.2f} MB'.format(end_mem))
    print('Decreased by {:.1f}%'.format(100 * (start_mem - end_mem) / start_mem))
    
    return df


def import_data(file):
    """create a dataframe and optimize its memory usage"""
    df = pd.read_csv(file, parse_dates=True, keep_date_col=True)
    df = reduce_mem_usage(df)
    return df

# Reading the bureau dataset
df_bureau = import_data('New_bureau.csv')

# Reading the bureau_balance dataset
df_bureau_bal = import_data('bureau_balance.csv')

df_bureau.head()

# Visualizing the info of the bureau dataframe
df_bureau.info()

# Visualizing the 'CREDIT_CURRENCY' categorical column of df_bureau
df_bureau['CREDIT_CURRENCY'].value_counts()

# Almost all of the rows have currency 1 as its value. Hence remove this column as it lacks variance
df_bureau.drop(columns='CREDIT_CURRENCY', inplace=True)

# Calculating the no. of previous loans information with credit bureaus for each applicant.
df_bureau['Prev_Bu_Loan_Cnt'] = df_bureau.groupby('SK_ID_CURR')['SK_ID_BUREAU'].transform('count')

df_bureau.head(5)

### Processing the bureau_bal dataframe

df_bureau_bal.head()

# Visualizing the info of df_bureau_bal dataframe
df_bureau_bal.info()

# Visualizing the unique values of the df_bureau_bal['STATUS'] categorical column
df_bureau_bal['STATUS'].value_counts()

# Getting the dummies for categorical column of df_bureau_bal
df_bureau_bal = pd.get_dummies(df_bureau_bal)

# Collapsing the df_bureau_bal dataframe on 'SK_ID_BUREAU' column
df_bureau_bal=df_bureau_bal.groupby('SK_ID_BUREAU',as_index=False).mean()

df_bureau_bal.head()

# Dropping Columns from df_bureau_bal to prevent linear dependency
df_bureau_bal.drop(columns=['STATUS_4','STATUS_5','STATUS_2','STATUS_3','STATUS_1'],inplace=True)

# Pre-pending bu_bal to each column name of dataframe df_bureau_bal
ls = ['bu_bal_'+col for col in df_bureau_bal.columns if col!='SK_ID_BUREAU']
ls.insert(0,'SK_ID_BUREAU')
df_bureau_bal.columns=ls

# Visualizing the result
df_bureau_bal.head()

# Getting the info of df_bureau_bal before merging
df_bureau_bal.info()

### Processing the bureau dataframe

# Visualizing the head of df_bureau
df_bureau.head()

# Visualizing the unique values of SK_ID_BUREAU column
df_bureau['SK_ID_BUREAU'].value_counts().head()

# Visualizing the unique values of SK_ID_CURR column
df_bureau['SK_ID_CURR'].value_counts().head()

# Visualizing the unique values of SK_ID_BUREAU column of df_bureau_bal
df_bureau_bal['SK_ID_BUREAU'].value_counts().head()

# Pre-pending bu to each column name of dataframe df_bureau
ls = ['bu_'+col for col in df_bureau.columns if col not in ['SK_ID_BUREAU','SK_ID_CURR']]
ls.insert(0,'SK_ID_CURR')
ls.insert(1,'SK_ID_BUREAU')
df_bureau.columns=ls

#Examining the results
df_bureau.head()

### Merging dataframes df_bureau & df_bureau_bal & analyzing the merged dataframes

# Merging df_bureau & df_bureau_bal on 'SK_ID_BUREAU' Column 
bu_bal_merged = df_bureau.merge(df_bureau_bal, on='SK_ID_BUREAU', how='left')

bu_bal_merged.head()

# Getting the Categorical columns of the merged dataframe bu_bal_merged
bu_bal_merged.info()

# Checking the unique values of bu_CREDIT_ACTIVE
bu_bal_merged['bu_CREDIT_ACTIVE'].value_counts()

# Checking the unique values of bu_CREDIT_TYPE
bu_bal_merged['bu_CREDIT_TYPE'].value_counts()

# Reclassifying minority values in 'bu_CREDIT_TYPE' as other
ls = ['Consumer credit','Credit card']
bu_bal_merged['bu_CREDIT_TYPE'] = bu_bal_merged['bu_CREDIT_TYPE'].astype(str)
bu_bal_merged.loc[~(bu_bal_merged['bu_CREDIT_TYPE'].isin(ls))
             ,'bu_CREDIT_TYPE'] = 'other'
bu_bal_merged['bu_CREDIT_TYPE'] = bu_bal_merged['bu_CREDIT_TYPE'].astype('category')

# verifing the result
bu_bal_merged['bu_CREDIT_TYPE'].value_counts()

# Getting dummies for the merged Dataframes
bu_bal_merged=pd.get_dummies(bu_bal_merged)

bu_bal_merged.head()

# Dropping redundant SK_ID_BUREAU column from the merged dataframe bu_bal_merged
bu_bal_merged.drop(columns='SK_ID_BUREAU',inplace=True)

bu_bal_merged.head()

# Cpllapsing the bu_bal_merged dataframe on 'SK_ID_CURR'
bu_bal_final=bu_bal_merged.groupby('SK_ID_CURR',as_index=False).mean()

#verifying the result
bu_bal_final.head()

# Redundant columns to drop from bu_bal_final to remove linear dependence
cols_list = ['bu_CREDIT_ACTIVE_Bad debt','bu_CREDIT_ACTIVE_Sold','bu_CREDIT_TYPE_other']

#Dropping Redundant columns from the bu_bal_final
bu_bal_final.drop(columns=cols_list,inplace=True)

bu_bal_final.head()

# Getting the info of the final merged dataframe
bu_bal_final.info()

# Saving the final merged dataframe into a file.
bu_bal_final.to_csv('bureau&bu_bal_final.csv',index=False)
