==> 15_Modeling.ipynb <==
# Importing all the relevant packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)# To see all the columns of a dataframe
#pd.set_option('display.max_rows', None)

# Function to reduce the memory usage of various Dataframes
def reduce_mem_usage(df):
    """ iterate through all the columns of a dataframe and modify the data type
        to reduce memory usage.        
       
        1. Iterate over every column
        2. Determine if the column is numeric
        3. Determine if the column can be represented by an integer
        4. Find the min and the max value
        5. Determine and apply the smallest datatype that can fit the range of values

    """
    start_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage of dataframe is {:.2f} MB'.format(start_mem))
    
    for col in df.columns:
        col_type = df[col].dtype
        
        if col_type != object:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
        else:
            df[col] = df[col].astype('category')

    end_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage after optimization is: {:.2f} MB'.format(end_mem))
    print('Decreased by {:.1f}%'.format(100 * (start_mem - end_mem) / start_mem))
    
    return df


def import_data(file):
    """create a dataframe and optimize its memory usage"""
    df = pd.read_csv(file, parse_dates=True, keep_date_col=True)
    df = reduce_mem_usage(df)
    return df

# Loadiing reduced feature training set
X_train = import_data('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loadiing reduced feature test set
X_test = import_data('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(5,random_state=42)

## Model_2  : Linear Discriminant Analysis Classifiers

### Computing the test set roc_auc score using Vanilla LDA calssifier with SVD solver with no other hyperparameters to tune.

# Importing LDA Classifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

# Instantiating the classifier object 
lda_svd = LinearDiscriminantAnalysis(solver='svd')

# Fitting the classifier on the Training data directly as there are no hyperparameters to tune
lda_svd.fit(X_train,y_train)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} is '.format(f_set, t_set, model_name), roc_auc_score(y, y_pred[:,1]))

# Calculating the Reduced feature training set roc_auc score.
cal_roc_auc(X_train, y_train, lda_svd, 'Reduced feature', 'training', 'LDA_SVD Classifier')

# Calculating the Reduced feature test set roc_auc score 
cal_roc_auc(X_test, y_test, lda_svd, 'Reduced feature', 'test', 'LDA_SVD Classifier')

### Computing the best hyperparameters for the Linear Discriminant Classifier with ‘eigen’ solver using Reduced Feature Training Set.

# Instantiating the classifier object 
lda_eigen = LinearDiscriminantAnalysis(solver='eigen')

# Defining the appropriate objective function for the LDA classifier
def objective_wrappper_lda(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Linear Discriminant Analysis parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'shrinkage': trial.suggest_uniform('shrinkage', 0, 1)
        }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def study_best_score_params(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Extracting the best model parameters and best study score
best_study_score,best_study_params = study_best_score_params(X_train, y_train, lda_eigen, objective_wrappper_lda, cv_strat, n_trials=200)

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best LDA  model with eigen solver by setting best study parameters.
lda_eigen = lda_eigen.set_params(**best_study_params)

# fitting the best LDA_eigen  model on the reduced feature training set
lda_eigen.fit(X_train, y_train)

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train, y_train, lda_eigen, 'Reduced feature', 'training', 'LDA Classifier with eigen solver & tuned shrinkage')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test, y_test, lda_eigen, 'Reduced feature', 'test', 'LDA Classifier with eigen solver & tuned shrinkage')

###  Computing the test set roc_auc score using LDA classifier with eigen solver & shrinkage utilizing the Ledoit-Wolf lemma with no other hyperparameters to tune.

# Instantiating the LDA classifier object using  eigen solver utilizing shrinkage using the Ledoit-Wolf lemma
lda_eigen_auto = LinearDiscriminantAnalysis(solver='eigen',shrinkage='auto')

# fitting the best LDA_eigen model with Ledoit-Wolf shrinkage on the reduced feature training set
lda_eigen_auto.fit(X_train, y_train)

# Calculating the Reduced feature training set roc_auc score using Ledoit-Wolf shrinkage
cal_roc_auc(X_train , y_train, lda_eigen_auto, 'Reduced feature', 'training', 'LDA Classifier with Ledoit-Wolf shrinkage')

# Calculating the Reduced feature test set roc_auc score using Ledoit-Wolf shrinkage
cal_roc_auc( X_test, y_test, lda_eigen_auto, 'Reduced feature', 'test', 'LDA Classifier with Ledoit-Wolf shrinkage')

### Calculating R_R ratio for Vanilla LDA Classifier using SVD solver

# Computing the CV scores using sklearn's cross_val_score
score_lda = cross_val_score(lda_svd, X_train, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the vanilla lda Classifier using roc_auc metric is: ',np.mean(score_lda))

print('The risk associated with the vanilla lda Classifier using roc_auc metric is: ',np.std(score_lda))

R_R_Ratio_lda = np.mean(score_lda)/np.std(score_lda)

print('The reward risk ratio for the vanilla lda Classifier using roc_auc metric is: ',R_R_Ratio_lda)

#### The R_R ratio for the vanilla lda Classifier using roc_auc metric is:  143.64202545415918

### Calculating R_R ratio for the best  lda Classifier using eigen solver & tuned shrinkage

# Computing the CV scores using sklearn's cross_val_score
score_lda_eigen = cross_val_score(lda_eigen, X_train, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best lda classifier with eigen solver using roc_auc metric is: ',np.mean(score_lda_eigen))

print('The risk associated with the best lda classifier with eigen solver using roc_auc metric is: ',np.std(score_lda_eigen))

R_R_Ratio_lda_eigen = np.mean(score_lda_eigen)/np.std(score_lda_eigen)

print('The reward risk ratio for the  best lda classifier with eigen solver using roc_auc metric is: ',R_R_Ratio_lda_eigen)

#### The R_R ratio for the best lda classifier with eigen solver & tuned shrinkage using roc_auc metric is:  142.3418843832703

### Calculating R_R ratio for the lda Classifier with eigen solver & shrinkage utilizing the Ledoit-Wolf lemma

# Computing the CV scores using sklearn's cross_val_score
score_lda_auto = cross_val_score(lda_eigen_auto, X_train, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: ',np.mean(score_lda_auto))

print('The risk associated with the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: ',np.std(score_lda_auto))

R_R_Ratio_lda_auto = np.mean(score_lda_auto)/np.std(score_lda_auto)

print('The reward risk ratio for the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: ',R_R_Ratio_lda_auto)

# Saving Various LDA models
import joblib
joblib.dump(lda_svd,'Vanilla_lda.joblib')
joblib.dump(lda_eigen,'Lda_eigen.joblib')
joblib.dump(lda_eigen_auto,'Lda_lw_lemma.joblib')

#### The R_R ratio for the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: 144.3922389528402

### R_R Ratio for the best LDA classifier using reduced feature set is:  144.3922389528402, corresponding to  LDA Classifier, utilizing LW lemma Shrinkage.

## Observations:
### 1) All the LDA variants have fitted the dataset pretty well, with no signs of overfitting & test set roc_auc score were well above that of the baseline model.
### 2) The  LDA model with highest test set roc_auc score corresponds to LDA with eigen solver & tuned shrinkage. This model  performed  worse than the tuned Logistic Regression Model on the test set. This was expected as the underlying feature space is not multivariate normal and doesn't have the same covariance matrix for both the classes ,which are the underlying assumption of the LDA models. Further the R_R ratio of the tuned Logistic Regression model was greater that of the LDA classifier using LW lemma Shrinkage.
### 3) As discussed in point 2 , for this dataset, the Logistic Regression model completely dominates all the 3 LDA variants in all respects.

## Model_3 : Quadratic Discriminant Analysis Classifier

### Computing the Reduced Feature test set roc_auc score using QDA classifier 

# Loading reduced feature training set again the usual way as QDA can't handle any data in float16 format
X_train = pd.read_csv('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading reduced feature test set again the usual way as QDA can't handle any data in float16 format
X_test = pd.read_csv('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis

# Instantiating the QDA classifier object 
QDA = QuadraticDiscriminantAnalysis()

# Defining the appropriate objective function for the QDA classifier
def objective_wrappper_qda(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Quadratic Discriminant Analysis parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'reg_param': trial.suggest_uniform('reg_param', 0, 1)
        }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def study_best_score_params(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Extracting the best model parameters and best study score
best_study_score,best_study_params = study_best_score_params(X_train, y_train, QDA, objective_wrappper_qda, cv_strat, n_trials=200)

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best reduced feature QDA  model with eigen solver by setting best study parameters.
QDA = QDA.set_params(**best_study_params)

# fitting the best QDA  model on the reduced feature training set
QDA.fit(X_train,y_train)

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train , y_train, QDA, 'Reduced feature', 'training', 'QDA classifier')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test, y_test, QDA, 'Reduced feature', 'test', 'QDA classifier')

# Saving the Reduced feature best QDA model 
import joblib
joblib.dump(QDA,'QDA.joblib')

### Calculating R_R ratio for best QDA Classifier.

# Computing the CV scores using sklearn's cross_val_score
score_qda = cross_val_score(QDA, X_train, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best QDA Classifier using roc_auc metric is: ',np.mean(score_qda))

print('The risk associated with the best QDA Classifier using roc_auc metric is: ',np.std(score_qda))

R_R_Ratio_qda = np.mean(score_qda)/np.std(score_qda)

print('The reward risk ratio for the best QDA Classifier using roc_auc metric is: ',R_R_Ratio_qda)

#### R_R Ratio for the best QDA classifier using roc_auc metric is:  118.95544640930348

## Observations:
### 1) The tuned QDA model has the worst test set roc_auc score as well as R_R ratio of all the fitted models till now. This was expected as the underlying Feature space is not multivariate normal, which is the underlying assumption of the QDA model. Any departure from normality affects the QDA more than the LDA, which is also observed here.
### 2) Owing to both very low  R_R ratio and test set roc_auc, we can rule out the QDA model for this dataset.

### R_R Ratio for the best classifier  in Discriminant Analysis family utilizing  roc_auc metric is:   144.3922389528402, corresponding to LDA Classifier, with LW lemma Shrinkage.

## Thus keeping everything into account, the best discriminant classifier is Vanilla LDA classifier  with SVD solver for this dataset.
