==> 16_Modeling.ipynb <==
# Importing Relevant Packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)# To see all the columns of a dataframe
#pd.set_option('display.max_rows', None)

# Function to reduce the memory usage of various Dataframes
def reduce_mem_usage(df):
    """ iterate through all the columns of a dataframe and modify the data type
        to reduce memory usage.        
       
        1. Iterate over every column
        2. Determine if the column is numeric
        3. Determine if the column can be represented by an integer
        4. Find the min and the max value
        5. Determine and apply the smallest datatype that can fit the range of values

    """
    start_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage of dataframe is {:.2f} MB'.format(start_mem))
    
    for col in df.columns:
        col_type = df[col].dtype
        
        if col_type != object:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
        else:
            df[col] = df[col].astype('category')

    end_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage after optimization is: {:.2f} MB'.format(end_mem))
    print('Decreased by {:.1f}%'.format(100 * (start_mem - end_mem) / start_mem))
    
    return df


def import_data(file):
    """create a dataframe and optimize its memory usage"""
    df = pd.read_csv(file, parse_dates=True, keep_date_col=True)
    df = reduce_mem_usage(df)
    return df

# Loading reduced feature training set
X_train = import_data('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading reduced feature test set
X_test = import_data('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

## Model_4: Light Gbm classifiers with Tuned Hyperparameters using Optuna.

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(5,random_state=42)

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Importing lightgbm Classifier
from lightgbm import LGBMClassifier

# defining Class weights
cl_weight = [None,'balanced',{0:1.0,1:12},{0:1.0,1:13},{0:1.0,1:14},{0:1.0,1:15},{0:1.0,1:16},
            {0:1.0,1:17},{0:1.0,1:18},{0:1.0,1:19},{0:1.0,1:20}]

# Defining the appropriate objective function for the Light GBM classifier

def objective_wrappper_lgbm(X_tr, y_tr, cls=None, cv_strat=None, cl_w=None):
    '''
    Optimizes classifier's cls (LightGBM here) parameters on the given training set X_tr, y_tr
    using cross-validation cv_strat & Class weights cl_w objects.
    
    '''
        
    def objective(trial):
        params = {
        'class_weight': trial.suggest_categorical('class_weight',cl_w),
        'n_estimators': trial.suggest_int('n_estimators', 10, 2000),
        'min_child_samples': trial.suggest_int('min_child_samples', 10, 250),
        'reg_alpha': trial.suggest_uniform('reg_alpha', 0, 10),
        'reg_lambda': trial.suggest_uniform('reg_lambda', 0, 10),
        'num_leaves': trial.suggest_int('num_leaves', 2, 250),
        'max_depth': trial.suggest_int('max_depth', 2, 100),
        'colsample_bytree': trial.suggest_uniform('colsample_bytree', 0.6, 1.0),
        'learning_rate': trial.suggest_loguniform('learning_rate', 1e-3, 1e-1)
        }
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def study_best_score_params(X_tr, y_tr, cls, obj_func, cv_strat, cl_w, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat, cl_w), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)

# Initializing the LightGBM Classifier
lgb_c = LGBMClassifier(random_state=42,objective='binary',n_jobs=5)

#### Computing the best hyperparameters for the LightGBM classifier_1 using Reduced Feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = study_best_score_params(X_train.values, y_train.values, lgb_c, objective_wrappper_lgbm, cv_strat,
                                                            cl_weight, n_trials=50)
# Used X_train.values & y_train.values as the some feature names had JSON characters , which were causing trouble while
# code execution.

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best tuned LightGbm model by setting best study parameters.
lgb_c = lgb_c.set_params(**best_study_params)

# fitting the best tuned lightgbm model on the Reduced feature training set.
lgb_c.fit(X_train.values, y_train.values)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} is '.format(f_set,t_set,model_name),roc_auc_score(y,y_pred[:,1]))

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train.values, y_train, lgb_c, 'Reduced feature', 'training', 'Light Gbm')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test.values, y_test, lgb_c, 'Reduced feature', 'test','Light Gbm')

### Calculating R_R ratio for the Light GBM Classifier_1.

# Computing the CV scores using sklearn's cross_val_score
score_Light_Gbm_1 = cross_val_score(lgb_c, X_train.values, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the Light GMB Classifier_1 using roc_auc metric is: ',np.mean(score_Light_Gbm_1))

print('The risk associated with the Light GMB Classifier_1 using roc_auc metric is: ',np.std(score_Light_Gbm_1))

R_R_Ratio_Light_Gbm_1 = np.mean(score_Light_Gbm_1)/np.std(score_Light_Gbm_1)

print('The reward risk ratio for the best Light Gbm Classifier using roc_auc metric is: ',R_R_Ratio_Light_Gbm_1)

print('5 fold CV roc_auc scores for the light_gbm classifier are: ',score_Light_Gbm_1)

# Saving the Reduced feature set best Light Gbm Classifier_1
import joblib
joblib.dump(lgb_c,'Light_Gbm_1.joblib')

#### R_R Ratio for the Light Gbm classifier_1 using roc_auc metric is:  166.50499771635376

## Observation(s):
### 1) Of all the models fitted till now, lgb classifier_1 has the best test set roc_auc score as well as the R_R ratio. 
### 2) But clearly lgb classifier_1 has overfitted the dataset , as there is a large difference between the training set and test set roc_auc score. This overfitting is typical of most boosting classifiers.

### Light Gbm Classifier_2 : A more regularized version of the previous light gbm model. 

# Defining the appropriate objective function for the more regularized Light GBM classifier

def objective_wrappper_lgbm(X_tr, y_tr, cls=None, cv_strat=None, cl_w=None):
    '''
    Optimizes classifier's cls (LightGBM here) parameters on the given training set X_tr, y_tr
    using cross-validation cv_strat & Class weights cl_w objects.
        
    '''
        
    def objective(trial):
        params = {
        'class_weight': trial.suggest_categorical('class_weight',cl_w),
        'n_estimators': trial.suggest_int('n_estimators', 10, 1200),# More Regularization
        'min_child_samples': trial.suggest_int('min_child_samples', 150, 250),# More Regularization
        'reg_alpha': trial.suggest_uniform('reg_alpha', 10, 15), # More Regularization
        'reg_lambda': trial.suggest_uniform('reg_lambda', 10, 15), # More Regularization
        'num_leaves': trial.suggest_int('num_leaves', 2, 175),# More Regularization
        'max_depth': trial.suggest_int('max_depth', 2, 75), # More Regularization
        'colsample_bytree': trial.suggest_uniform('colsample_bytree', 0.6, 1.0),
        'learning_rate': trial.suggest_loguniform('learning_rate', 1e-3, 1e-1)
        }
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def study_best_score_params(X_tr, y_tr, cls, obj_func, cv_strat, cl_w, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat, cl_w), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)

# Initializing the LightGBM Classifier
lgb_c = LGBMClassifier(random_state=42, objective='binary', n_jobs=5)

#### Computing the best hyperparameters for the Light GBM classifier_2 using Reduced Feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = study_best_score_params(X_train.values, y_train.values, lgb_c, objective_wrappper_lgbm, cv_strat,
                                                            cl_weight, n_trials=50)
# Used X_train.values & y_train.values as the some feature names had JSON characters , which were causing trouble 
# during execution.

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best tuned & more regularized LightGbm model by setting best study parameters.
lgb_c = lgb_c.set_params(**best_study_params)

# fitting the best tuned  & more regularized lightgbm model on the Reduced feature training set
lgb_c.fit(X_train.values, y_train.values)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} is '.format(f_set,t_set,model_name),roc_auc_score(y,y_pred[:,1]))

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train.values, y_train, lgb_c, 'Reduced feature', 'training', 'updated Light Gbm')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test.values, y_test, lgb_c, 'Reduced feature', 'test','updated Light Gbm')

### Calculating R_R ratio for the Light GBM Classifier_2.

# Computing the CV scores using sklearn's cross_val_score for more regularized light gbm classifier_2
score_Light_Gbm_2 = cross_val_score(lgb_c, X_train.values, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the Light GMB Classifier_2 using roc_auc metric is: ',np.mean(score_Light_Gbm_2))

print('The risk associated with the Light GMB Classifier_2 using roc_auc metric is: ',np.std(score_Light_Gbm_2))

R_R_Ratio_Light_Gbm_2 = np.mean(score_Light_Gbm_2)/np.std(score_Light_Gbm_2)

print('The reward risk ratio for the best Light Gbm Classifier_2 using roc_auc metric is: ',R_R_Ratio_Light_Gbm_2)

print('5 fold CV roc_auc scores for the light_gbm classifier_2 are: ',score_Light_Gbm_2)

# Saving the Reduced feature set best updated regularized  Light Gbm Classifier 
import joblib
joblib.dump(lgb_c,'Light_Gbm_2.joblib')

#### R_R Ratio for the more regularized Light Gbm classifier_2 using roc_auc metric is:  167.37634851965672

## Observations:
### 1) By making use of more regularization, Light Gbm classifier_2 has been able to substantially reduce overfitting as compared to that of Light Gbm classifier_1, with the test set roc_auc scores being almost equal for both of them.
### 2) Further R_R ratio for the Light Gbm classifier_2 is marginally higher than that of Light Gbm classifier_1.
### 3) Thus Light Gbm classifier_2 beats the Light Gbm classifier_1 hands down and has the best test set roc_auc as well as R_R ratio of all the models tested till now for this dataset.

### R_R Ratio for the best  Light Gbm classifier using roc_auc metric is:  167.37634851965672 , corresponding to more regularized Light Gbm classifier_2.
