==> 19_Modeling.ipynb <==
# importing Required Packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)# To see all the columns of a dataframe
#pd.set_option('display.max_rows', None)

# Function to reduce the memory usage of various Dataframes
def reduce_mem_usage(df):
    """ iterate through all the columns of a dataframe and modify the data type
        to reduce memory usage.        
       
        1. Iterate over every column
        2. Determine if the column is numeric
        3. Determine if the column can be represented by an integer
        4. Find the min and the max value
        5. Determine and apply the smallest datatype that can fit the range of values

    """
    start_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage of dataframe is {:.2f} MB'.format(start_mem))
    
    for col in df.columns:
        col_type = df[col].dtype
        
        if col_type != object:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
        else:
            df[col] = df[col].astype('category')

    end_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage after optimization is: {:.2f} MB'.format(end_mem))
    print('Decreased by {:.1f}%'.format(100 * (start_mem - end_mem) / start_mem))
    
    return df


def import_data(file):
    """create a dataframe and optimize its memory usage"""
    df = pd.read_csv(file, parse_dates=True, keep_date_col=True)
    df = reduce_mem_usage(df)
    return df

# Loading reduced feature training set
X_train = import_data('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading reduced feature test set
X_test = import_data('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

## Loading all the best models from various categories

# Importing Joblib module
import joblib

# Importing best Logistic regression Classifier
lr = joblib.load('Log_Reg.joblib')

# Importing best Random Forest Classifier
rf = joblib.load('Rand_Forest_optuna_reg_2.joblib')

# Importing best Light Gbm Classifier
lgbm = joblib.load('Light_Gbm_2.joblib')

# Importing best Linear Discriminant Analysis Classifier
lda = joblib.load('Vanilla_lda.joblib')

## Model_9: Voting Classifier with Default Weights & Soft_Voting.

# Importing Voting classifier from sklearn
from sklearn.ensemble import VotingClassifier

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Instantiating the voting Classifier Object
voting_clf = VotingClassifier(estimators=[('logistic_Reg',lr),('Random_Forest',rf),('Light_Gbm',lgbm),
                                           ('Linear_Dis',lda)],voting='soft',n_jobs=5

# Fitting the default voting classifier on the Reduced Feature Training set
voting_clf.fit(X_train.values, y_train)

#Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the {} is '.format(f_set,t_set,model_name),roc_auc_score(y,y_pred[:,1]))

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train.values, y_train, voting_clf,'reduced feature','training',' default Voting Classifier')

# Calculating the reduced feature test roc_auc score using the best study parameters
cal_roc_auc(X_test.values, y_test, voting_clf,'reduced feature','test','default Voting Classifier')

# Saving the default voting Classifier
import joblib
joblib.dump(voting_clf,'Voting_default.joblib')

### Calculating R_R ratio for default Voting Classifier.

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(5,random_state=42)

# Computing the CV scores using sklearn's cross_val_score
score_voting_default = cross_val_score(voting_clf, X_train.values, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the default Voting Classifier using roc_auc metric is: ',np.mean(score_voting_default))

print('The risk associated with the default Voting Classifier using roc_auc metric is: ',np.std(score_voting_default))

R_R_Ratio_voting_default = np.mean(score_voting_default)/np.std(score_voting_default)

print('The reward risk ratio for the default Voting Classifier using roc_auc metric is: ',R_R_Ratio_voting_default)

# Saving the default voting Classifier
import joblib
joblib.dump(voting_clf,'Voting_default.joblib')

#### R_R Ratio for the default Voting classifier using roc_auc metric is: 161.03129442118123

## Observation(s): 
### 1 The default voting classifier slightly overfits the dataset. May be hyperparameter tuning give better results.
### 2) The test set roc_auc score for the default voting classifier is more than that of component Logistic Regression , LDA & Random Forest models, but less than that of  component Light GBM. May be tuning the weights of voting classifier result in better results.
### 3) On the other hand, the R_R ratio for the default voting classifier model is more than those of component Logistic regression &  LDA models, but less than those of component Random forest & LightGBM models. So clearly LightGBM bests default Voting classifier hands down.

## Model_10: Voting Classifier with Tuned weights using Optuna.

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Instantiating a new voting classifier object
voting_clf_2 = VotingClassifier(estimators=[('logistic_Reg',lr),('Random_Forest',rf),('Light_Gbm',lgbm),
                                           ('Linear_Dis',lda)],voting='soft',n_jobs=5)

# Defining the appropriate objective function for the best weights of Voting classifier.
def objective_wrappper_Vt_2(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes voting classifier (cls) weights parameter on the given training set X_tr,y_tr
    using Startified Cross Validation object cv_strat
      
    '''
    
    def objective(trial):
        w1 = trial.suggest_uniform('w1',-1,1)
        w2 = trial.suggest_uniform('w2',-1,1)
        w3 = trial.suggest_uniform('w3',-1,1)
        w4 = trial.suggest_uniform('w4',-1,1)
        
        params = {
            'weights':[w1,w2,w3,w4]                 
            }
        
        cls.set_params(**params)#Initializing the model with the parameter
               
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))
    
    return objective

#### Computing the best weights for the Voting Classifier_2 using Optuna.

# Defining the evaluation function for study's best parameters
def study_best_score_params(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=50):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Extracting the best model parameters and best study score for the Voting Classifier_2
best_study_score,best_study_params = study_best_score_params(X_train.values, y_train, voting_clf_2, objective_wrappper_Vt_2,
                                                   cv_strat)

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best tuned Voting clasiifier model by setting best study parameters.
voting_clf_2 = voting_clf_2.set_params(weights=[best_study_params['w1'],best_study_params['w2'],
                                                 best_study_params['w3'],best_study_params['w4']])

# fitting the best Voting Classifier on the reduced feature training set
voting_clf_2.fit(X_train.values, y_train)

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train.values, y_train, voting_clf_2, 'reduced feature', 'training', 'tuned Voting Classifier_2')

# Calculating the reduced feature test roc_auc score using the best study parameters
cal_roc_auc(X_test.values, y_test, voting_clf_2, 'reduced feature', 'test', 'tuned Voting Classifier_2')

# Saving the tuned voting Classifier
import joblib
joblib.dump(voting_clf_2,'Voting_class_2.joblib')

### Calculating R_R ratio for  tuned  Voting Classifier_2.

# Computing the CV scores using sklearn's cross_val_score
score_voting_2 = cross_val_score(voting_clf_2, X_train.values, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the tuned Voting Classifier using roc_auc metric is: ',np.mean(score_voting_2))

print('The risk associated with the tuned Voting Classifier using roc_auc metric is: ',np.std(score_voting_2))

R_R_Ratio_voting_2 = np.mean(score_voting_2)/np.std(score_voting_2)

print('The reward risk ratio for the tuned Voting Classifier_2 using roc_auc metric is: ',R_R_Ratio_voting_2)

#### R_R Ratio for the tuned Voting classifier_2 using roc_auc metric is: 163.79317380910297

## Observation(s): 
### 1 The tuned voting classifier_2  clearly overfits the dataset, even more than the default version.
### 2) The test set roc_auc score of the tuned voting classifier_2 is more than that of all component models (by a good margin ), but for that of the component Light  GBM and also exceeds that of untuned  default version.
### 3) Further, the R_R ratio for the tuned voting classifier model is more than that of the default voting classifier, but still less than those of component Light Gbm and Random Forest classifiers.

#### Lets try to further reduce overfitting by incorporating more regularization & tuning another Voting Classifier_3.

# Instantiating a new voting classifier object
voting_clf_3 = VotingClassifier(estimators=[('logistic_Reg',lr),('Random_Forest',rf),('Light_Gbm',lgbm),
                                           ('Linear_Dis',lda)],voting='soft',n_jobs=5)

# Defining the appropriate objective function for the best positive weights of Voting classifier.
def objective_wrappper_Vt_3(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes voting classifier (cls) weights parameter on the given training set X_tr,y_tr
    using Startified Cross Validation object cv_strat
      
    '''
    
    def objective(trial):
        w1 = trial.suggest_uniform('w1',0,1)
        w2 = trial.suggest_uniform('w2',0,1)
        w3 = trial.suggest_uniform('w3',0,1)
        w4 = trial.suggest_uniform('w4',0,1)
        
        params = {
            'weights':[w1,w2,w3,w4]                 
            }
        
        cls.set_params(**params)#Initializing the model with the parameter
               
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))
    
    return objective

# Extracting the best model parameters and best study score for the 2nd Voting Classifier
best_study_score,best_study_params = study_best_score_params(X_train.values, y_train, voting_clf_3, objective_wrappper_Vt_3,
                                                   cv_strat)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ',best_study_params))

# Obtaining the best tuned voting clasiifier model by setting best study parameters.
voting_clf_3 = voting_clf_3.set_params(weights=[best_study_params['w1'],best_study_params['w2'],
                                                 best_study_params['w3'],best_study_params['w4']])

# Fitting the best tuned Voting Classifier_3 on the reduced feature training set.
voting_clf_3.fit(X_train.values, y_train)

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train.values, y_train, voting_clf_3, 'reduced feature', 'training', 'tuned Voting Classifier_3')

# Calculating the reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test.values, y_test, voting_clf_3, 'reduced feature', 'test', 'tuned Voting Classifier_3')

# Saving the best voting Classifier_3
import joblib
joblib.dump(voting_clf_3,'Voting_class_3.joblib')

### Calculating R_R ratio for the tuned Voting Classifier_3.

# Computing the CV scores using sklearn's cross_val_score
score_voting_3 = cross_val_score(voting_clf_3, X_train.values, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the tuned Voting Classifier_3 using roc_auc metric is: ',np.mean(score_voting_3))

print('The risk associated with the tuned Voting Classifier_3 using roc_auc metric is: ',np.std(score_voting_3))

R_R_Ratio_voting_3 = np.mean(score_voting_3)/np.std(score_voting_3)

print('The reward risk ratio for the tuned Voting Classifier_3 using roc_auc metric is: ',R_R_Ratio_voting_3)

print('5 fold CV roc_auc scores for the tuned Voting Classifier_3 are: ',score_voting_3)

#### R_R Ratio for the tuned Voting classifier_3 using roc_auc metric is: 165.12186332349236

## Observation(s): 
### 1 The tuned voting classifier_3  clearly overfits the dataset, even slightly more than the voting classifier_2.
### 2) The test set roc_auc score of the tuned voting classifier_3 is again more than that of all component models (by a good margin ), but for that of the component Light  GBM and also exceeds those of both untuned  default version & Voting classifier_2
### 3) Further, the R_R ratio for the tuned voting classifier_3 model is more than those of both voting classifier_2 & the default version , but still less than that of the component Random Forest classifier.

### Best R_R Ratio for the voting classifier family using roc_auc metric is:  165.12186332349236 ,corresponding to tuned Voting Classifier_3.

### Taking everything into consideration, the tuned Voting Classifier_3 , even with slight overfitting, dominates all other voting classifiers & is clearly the best in the family of Voting Classifiers for this dataset.
